create a seperate component for tooltip
