+++ frontend/public/extend/devconsole/components/topology/shapes/VictoryPodStatus.tsx
@@ -1,9 +1,8 @@
 import * as React from 'react';
 import { VictoryPie } from 'victory';
 import './DefaultGroup.scss';
-import SvgDropShadowFilter from './../../../shared/components/svg/SvgDropShadowFilter';
-import { createFilterIdUrl } from './../../../shared/utils/svg-utils';
 import { Pod } from '../topology-types';
+import Tooltip from '../../../shared/components/svg/SvgPodTooltip';
 
 export const podColor = {
   Running: '#00b9e4',
@@ -105,79 +104,5 @@ export default class PodStatus extends React.PureComponent<PodStatusProps> {
         }}
       />
     );
-  }
-}
-
-type TooltipProps = {
-  datum?: any;
-  active?: boolean;
-  x: number;
-  y: number;
-};
-
-type State = {
-  bb: SVGRect;
-};
-
-const FILTER_ID = 'PodTOOLTIP';
-
-class Tooltip extends React.PureComponent<TooltipProps> {
-  state: State = {
-    bb: null,
-  };
-
-  groupRef = React.createRef<SVGGElement>();
-
-  componentDidMount() {
-    this.computeBoxSize();
-  }
-
-  componentDidUpdate({ datum }: TooltipProps) {
-    if (this.props.datum !== datum) {
-      this.computeBoxSize();
-    }
-  }
-
-  private computeBoxSize() {
-    if (this.props.active) {
-      this.setState({ bb: this.groupRef.current.getBBox() });
-    }
-  }
-
-  render() {
-    const { bb } = this.state;
-    const { x, y } = this.props;
-    const paddingX = 20;
-    const paddingY = 5;
-    return this.props.active ? (
-      <g className={'odc-default-group__label'}>
-        <SvgDropShadowFilter id={FILTER_ID} />
-        {bb && (
-          <React.Fragment>
-            <rect
-              filter={createFilterIdUrl(FILTER_ID)}
-              x={x - paddingX - bb.width / 2}
-              width={bb.width + paddingX * 2}
-              y={y - paddingY - bb.height / 2}
-              height={bb.height + paddingY * 2}
-            />
-            <rect
-              width={10}
-              height={10}
-              x={x - bb.width / 2 - paddingX / 2}
-              y={y - paddingY}
-              style={{ fill: podColor[this.props.datum.x] }}
-            />
-          </React.Fragment>
-        )}
-        <g ref={this.groupRef}>
-          <text x={x} y={y} dx={10} textAnchor="middle" dy="0.3em">
-            <tspan>{this.props.datum.x}</tspan>
-            <tspan dx={20}>{this.props.datum.pods}</tspan>
-          </text>
-        </g>
-      </g>
-    ) :
-    null;
   }
 }
